+++ app/javascript/users/components/homepage/cookie_policy_content.vue
@@ -1,0 +1,136 @@
+<template>
+  <div>
+    <h1 class="subtitle">Cookie Policy</h1>
+    <p class="subtitle text-center"> Ultimo aggiornamento: 20 settembre 2020 </p>
+
+    <p>{{ organizationName }}, titolare del trattamento, ti informa ai sensi della normativa nazionale applicabile e
+      dell’art. 13 Regolamento Privacy Europeo n. 679/2016 (“GDPR”) che questo sito web (“Sito”) fa uso di cookie.
+      Desideriamo che tu ti senta sempre sicuro e tutelato. Per questo ti invitiamo a leggere questa informativa,
+      dove ti spiegheremo cosa sono i cookie, come li utilizziamo e come li controlliamo.<br />
+      In questo documento desideriamo spiegarti:</p>
+    <ol>
+      <li>che cosa sono i cookie;</li>
+      <li>quali cookie vengono utilizzati;</li>
+      <li>a cosa servono i cookie;</li>
+      <li>come puoi controllare ed eliminare i cookie.</li>
+    </ol>
+
+    <h2>1. Che cosa sono i cookie?</h2>
+    <p>I cookie sono piccoli files di testo che vengono inviati sul tuo dispositivo elettronico (PC, tablet, smartphone, ecc.)
+      ogni volta che navighi su un sito web e che, durante le navigazioni successive, vengono rinviati al sito web
+      di origine, o inviati ad un diverso sito web che sia in grado di riconoscere quello specifico cookie.
+      I cookie memorizzano le interazioni del tuo dispositivo con il sito web. In questo modo il sito riconoscerà
+      il tuo device ad ogni visita successiva e ricorderà le tue preferenze di navigazione per offrirti un’esperienza
+      più funzionale e contenuti personalizzati.</p>
+
+    <h2>2. Quali cookie vengono utilizzati?</h2>
+    <p>I cookies possono essere tecnici, analitici e di profilazione.<br />
+      Il Sito utilizza cookie: <br />
+    </p>
+    <ol>
+      <li>tecnici, sia “di sessione”, ossia cookie che vengono cancellati alla chiusura del browser, sia persistenti
+        di prima parte, ossia cookie che scadono dopo un determinato periodo di tempo e che sono inviati direttamente
+        dal nostro sito.</li>
+      <br />
+      <li>analitici di prima parte, ossia i dati che il sito raccoglie in forma aggregata per sapere quanti utenti
+        visitano il sito e con quali modalità e analitici di terza parte, ossia condivisi con domini diversi rispetto
+        a quello indicato nella barra degli indirizzi. {{ organizationName }} non è responsabile per l’eventuale utilizzo delle informazioni
+        effettuato da terze parti, mediante i cookies sopra indicati. Questi cookie, infatti, non vengono controllati
+        direttamente da {{ organizationName }}, per revocare detto consenso è necessario fare riferimento ai siti internet delle terze parti.
+        Questi cookies sono:<br />
+        - Google Analytics: un servizio di analisi web fornito da Google Inc., 1600 Amphitheatre Parkway, Mountain View,
+        CA 94043, USA ("Google"). Google Analytics utilizza anche i cookie, che consentono di analizzare l'utilizzo
+        del portale da parte dell'utente. Le informazioni generate dal cookie
+        sull'utilizzo del portale vengono solitamente trasmesse e archiviate su un server di Google negli Stati Uniti.
+        L'anonimizzazione dell'IP è stata attivata su questo portale, in modo che l'indirizzo IP degli utenti di Google
+        all'interno degli Stati membri dell'Unione europea o in altri Stati contraenti dell'accordo sullo Spazio economico
+        europeo sia stato precedentemente abbreviato. Solo in casi eccezionali l'indirizzo IP completo verrà inviato
+        a un server di Google negli Stati Uniti e abbreviato. Per conto del Fornitore, Google utilizzerà queste informazioni
+        per valutare l'utilizzo del portale, compilare report sulle attività del sito Web e fornire altri servizi
+        relativi alle attività del sito Web e all'utilizzo di Internet al Provider. L'indirizzo IP fornito da Google Analytics
+        come parte di Google Analytics non verrà unito ad altri dati di Google. <br>
+        È possibile impedire la memorizzazione dei cookie attraverso un'impostazione del software del browser (indicato più sotto);
+        Tuttavia, il Fornitore ti informa che in questo caso potresti non essere in grado di utilizzare tutte le funzioni
+        di questo sito web nella misura massima. Inoltre, puoi impedire la raccolta da parte di Google dei dati generati
+        dal cookie e relativi al tuo utilizzo del portale (compreso il tuo indirizzo IP) nonché l'elaborazione di questi
+        dati da parte di Google scaricando e installando il plug-in del browser disponibile al seguente link:
+
+        https://tools.google.com/dlpage/gaoptout?hl=en.
+      </li>
+      <br />
+      <li>di profilazione di prima e di terza parte, ossia cookie che appartengono al Sito
+        o ad altri siti web e servizi e che ci aiutano a capire meglio i tuoi gusti e necessità.</li>
+    </ol>
+
+    <br />
+    <p>Se desideri saperne di più sui cookies installati sul tuo dispositivo e/o browser
+      e/o disabilitarli, clicca i seguenti link:<br />
+      - <a></a><a href="http://www.youronlinechoices.com/uk/your-ad-choices" target="_blank">http://www.youronlinechoices.com/uk/your-ad-choices</a><u><br />
+      </u>- <a href="http://www.allaboutcookies.org/manage-cookies/index.html">http://www.allaboutcookies.org/manage-cookies/index.html</a><br />
+      - <a href="https://priv-policy.imrworldwide.com/priv/browser/it/it/optout.html">https://priv-policy.imrworldwide.com/priv/browser/it/it/optout.html</a></p>
+    <p style="margin-left: 40.0px;"></p>
+
+    <h2>3. A cosa servono i cookie?</h2>
+    <p>I cookie tecnici usati dal nostro Sito servono a effettuare e facilitare la tua navigazione e a permetterti di
+      fruire dei servizi del Sito. Ad esempio, ti consentono di non dover inserire nuovamente dati per il login.
+      I cookie analitici servono ad analizzare e monitorare il modo in cui utilizzi il Sito (ad esempio il numero di
+      accessi e le pagine viste), a fini statistici e per permetterci di offrirti un’esperienza di navigazione sempre
+      più facile e funzionale. I cookie di profilazione servono a tracciare la tua navigazione sul Sito e a individuare
+      i tuoi gusti, abitudini, scelte, ecc. Questi cookie possono essere usati per trasmetterti messaggi pubblicitari
+      in linea con le preferenze che hai già manifestato nella navigazione online.</p>
+
+    <h2>4. Come puoi controllare ed eliminare i cookie?</h2>
+    <p>La prima volta che visiti il sito vedrai comparire un banner relativo all’utilizzo dei cookie. Potrai
+      quindi decidere se acconsentire oppure no all’uso dei cookie. Potrai comunque disattivarli in ogni momento,
+      cambiando la configurazione del tuo browser e seguendo le istruzioni contenute nelle policy dei gestori dei cookie. <br />
+      Per fare in modo che il tuo browser rigetti tutti o alcuni cookie, fai riferimento alle istruzioni fornite dal
+      provider del browser. In particolare: <br />
+    </p>
+    <ol>
+      <li>Microsoft Internet Explorer <br />
+        Cliccare l’icona 'Strumenti' nell’angolo in alto a destra e selezionare 'Opzioni internet'. Nella finestra
+        pop-up selezionare 'Privacy'. Qui gli Utenti possono modificare le impostazioni dei cookie.</li>
+      <li>Google Chrome <br />
+        Cliccare l’icona “chiave inglese” nell’angolo in alto a destra e selezionare 'Impostazioni'. A questo punto
+        selezionare 'Mostra impostazioni avanzate’ e cambiare le impostazioni della 'Privacy'.</li>
+      <li>Mozilla Firefox <br />
+        Dal menu a tendina nell’angolo in alto a sinistra selezionare 'Opzioni'. Nella finestra di pop-up selezionare
+        'Privacy'. Qui gli Utenti possono regolare le impostazioni dei cookie.</li>
+      <li>Safari <br />
+        Dal menu di impostazione a tendina nell’angolo in alto a destra selezionare 'Preferenze'. Selezionare
+        'Sicurezza' e qui gli Utenti possono regolare le impostazioni dei cookie. Trovi una guida maggiormente
+        dettagliata al link www.aboutcookie.org</li>
+    </ol>
+    <p>Per maggiori informazioni puoi leggere la nostra <v-btn text class="" :to='{name: "Privacy"}'>Informativa Privacy</v-btn>.</p>
+
+    <h2>Consenso</h2>
+    <p>Usando il nostro sito web, acconsenti alla nostra politica sull'utilizzo dei cookie e accetti i suoi termini. Se desideri
+      ulteriori informazioni o hai domande non esitare a contattarci.</p>
+  </div>
+</template>
+
+<script>
+  export default {
+    props: {
+      organizationName: {
+        type: String,
+        required: true,
+      },
+    },
+  }
+</script>
+
+<style lang="scss" scoped>
+  h1, h2, h3 {
+    text-align: center;
+    color: #666666;
+    margin-top: 2em;
+  }
+  p {
+    color: #444444;
+    &.subtitle {
+      text-align: center;
+      color: #888888;
+    }
+  }
+</style>